# swagger.yaml
openapi: 3.0.0
info: 
  title: Simple Inventory API
  description: A simple inventory API
  version: 1.0.0
  contact: {}
servers:
  - url: https://solid-meme.olalekantemitay.repl.co/api/v1
    variables: {}
paths:
  /items:
    post:
      tags:
        - Items
      summary: Create Item
      operationId: CreateItem
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
            example:
              name: pencil
              quantity: 1000
        required: true
      responses:
        '201':
          description: 'Item created successfully'
          headers: {}
        '400':
          description: 'The item name exists! or  the name or quantity is not in the request Body '
          headers: {}
        '500':
          description: 'Server Error'
          headers: {}
      deprecated: false
    get:
      tags:
        - Items
      summary: Get Items
      operationId: GetItems
      parameters: []
      responses:
        '200':
          description: 'Item fetched successfully'
          headers: {}
      deprecated: false
...
components:
  schemas:
    CreateItemRequest:
      title: CreateItemRequest
      required:
        - name
        - quantity
      type: object
      properties:
        name:
          type: string
        quantity:
          type: int
      example:
        name: Pencil
        quantity: 200
...
security:
  - httpBearer: []
...